fix merge conflicts

+++ server/routes/admin/upload.php
@@ -14,7 +14,7 @@
     return;
   }
 
-  $isValidAdmin = validAdmin($_POST["adminKey"], '../../database/ljep.db');
+  $isValidAdmin = validAdmin($_POST["adminKey"], '../../database/admin.db');
   if(!$isValidAdmin) return;
 
   $config = json_decode(file_get_contents(dirname(__DIR__, 2) . '/config.json'));
@@ -106,10 +106,18 @@
     }
   }
 
+  if (isDatabaseLocked()) {
+    http_response_code(503);
+    echo json_encode(array("error" => "Database is locked by another process. Please try again later. Note if you think the database is incorrectly locked then it will automatically unlock in 2 hours."));
+    return;
+  }
+  lockDatabase();
+
   $uploadStartTime = time();
   ingestSpreadsheet($newFilePath, basename($newFilePath), 0);
   deleteOldCrknData('filename', $uploadStartTime, basename($newFilePath));
 
+  unlockDatabase();
   $serverFiles = getXLSXFiles('../../PAR-files/');
   echo json_encode(array("files" => $serverFiles));
 ?>
